Extract DeleteProjectModal into shared component file
